ajout de relation de table pour les quacks appartenant au ducks, full front avec tailwind, et connexion, déconnexion, création de compte à jour

+++ src/Controller/QuackController.php
@@ -9,6 +9,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 #[Route('/quack')]
 class QuackController extends AbstractController
@@ -24,7 +25,10 @@ class QuackController extends AbstractController
     #[Route('/new', name: 'app_quack_new', methods: ['GET', 'POST'])]
     public function new(Request $request, QuackRepository $quackRepository): Response
     {
+        $user = $this->getUser();
+
         $quack = new Quack();
+        $quack->setUser($user);
         $form = $this->createForm(QuackType::class, $quack);
         $form->handleRequest($request);
 
@@ -51,6 +55,12 @@ class QuackController extends AbstractController
     #[Route('/{id}/edit', name: 'app_quack_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Quack $quack, QuackRepository $quackRepository): Response
     {
+
+        // Vérifie si l'utilisateur connecté est le propriétaire du quack
+        if ($this->getUser() !== $quack->getUser()) {
+            throw new AccessDeniedException('Vous n\'êtes pas autorisé à modifier ce quack.');
+        }
+
         $form = $this->createForm(QuackType::class, $quack);
         $form->handleRequest($request);
 
@@ -69,6 +79,11 @@ class QuackController extends AbstractController
     #[Route('/{id}', name: 'app_quack_delete', methods: ['POST'])]
     public function delete(Request $request, Quack $quack, QuackRepository $quackRepository): Response
     {
+        // Vérifie si l'utilisateur connecté est le propriétaire du quack
+        if ($this->getUser() !== $quack->getUser()) {
+            throw new AccessDeniedException('Vous n\'êtes pas autorisé à supprimer ce quack.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$quack->getId(), $request->request->get('_token'))) {
             $quackRepository->remove($quack, true);
         }
